swals!

+++ src/store/modules/CarStore.js
@@ -1,6 +1,6 @@
 import router from "../../router"
 import { api } from "../../services/AxiosService"
-
+import ns from "../../services/NotificationService"
 export default {
   actions: {
     async getAllCars({ commit }) {
@@ -37,6 +37,7 @@ export default {
       try {
         let res = await api.put('cars/' + bid.id, bid)
         commit("setResource", { resource: "activeCar", data: res.data })
+        ns.toast("Bid Increased", 2000)
       } catch (error) {
         console.error(error)
       }
@@ -44,11 +45,14 @@ export default {
     },
     async deleteCar({ commit }, id) {
       try {
-        await api.delete('cars/' + id)
-        commit("removeResource", { resource: "cars", id })
-        commit("setResource", { resource: "activeCar", data: {} })
-        // NOTE this will change the active route
-        router.push({ name: "Cars" })
+        if (await ns.confirmAction("Do ya wanna delete this car?", "Check yoself, b4 you wreck yoself")) {
+          await api.delete('cars/' + id)
+          commit("removeResource", { resource: "cars", id })
+          commit("setResource", { resource: "activeCar", data: {} })
+          // NOTE this will change the active route
+          router.push({ name: "Cars" })
+          ns.toast("Car Deleted", 3000, "warning")
+        }
       } catch (error) {
         console.error(error)
       }
